FirstHw email feedback tasks changes

+++ 2th_semester/Haskell/Homeworks/FirstLectureHw/src/Task6.hs
@@ -3,11 +3,11 @@ module Task6
   )
 where
   
-canCarry :: Double -> Double -> Double -> String
+canCarry :: Int -> Double -> Double -> String
 canCarry c k w 
 --  | [x | x <- [c, k, w], x < 0] /= [] = error "Error negative number given"
   | c < 0 = error "The number of products was negative"
   | k < 0 = error "John's hosting capacity was negative"
   | w < 0 = error "The weight of a product was negative"
-  | c * w <= k = "Yes"
+  | fromIntegral c * w <= k = "Yes"
   | otherwise = "No"
